requirements.txt
python-telegram-bot==13.7
requests
redis
urllib3==1.26.15